+++ homework08/src/main/java/ru/otus/spring/homework08/service/CommentServiceImpl.java
@@ -3,9 +3,11 @@ package ru.otus.spring.homework08.service;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
+import ru.otus.spring.homework08.domain.Book;
 import ru.otus.spring.homework08.domain.Comment;
 import ru.otus.spring.homework08.exceptions.IllegalBookException;
 import ru.otus.spring.homework08.exceptions.IllegalCommentException;
+import ru.otus.spring.homework08.storage.BookDao;
 import ru.otus.spring.homework08.storage.CommentDao;
 
 import java.util.List;
@@ -17,14 +19,14 @@ public class CommentServiceImpl implements CommentService {
 
     private final CommentDao commentDao;
     private final IOService ioService;
-    private final CustomService customService;
+    private final BookDao bookDao;
 
     @Override
     public String insert() {
         ioService.output("Укажите книгу, для который хотите написать комменатрий");
         String bookname = ioService.input();
         try {
-            Comment temp = new Comment(customService.checkBook(bookname));
+            Comment temp = new Comment(checkBook(bookname));
             ioService.output("Напишите ваш комментарий");
             String strComment = ioService.input();
             temp.setComment(strComment);
@@ -46,7 +48,7 @@ public class CommentServiceImpl implements CommentService {
     public void showById() {
         ioService.output("Показать комментарий с id");
         String id = ioService.input();
-        Comment temp = checkCommentById(id);
+        Comment temp = getCommentById(id);
         ioService.output(temp.toString());
     }
 
@@ -54,7 +56,7 @@ public class CommentServiceImpl implements CommentService {
     public void showByBook() {
         ioService.output("Показать комменатрий для книги с названием");
         String bookname = ioService.input();
-        List<Comment> list = checkCommentsByBookName(bookname);
+        List<Comment> list = getCommentsByBookName(bookname);
         list.forEach(temp -> ioService.output(temp.toString()));
     }
 
@@ -62,7 +64,9 @@ public class CommentServiceImpl implements CommentService {
     public String deleteById() {
         ioService.output("Удалить комментарий по id");
         String id = ioService.input();
-        checkCommentById(id);
+        if (commentDao.getById(id) == null) {
+            throw new IllegalCommentException();
+        }
         commentDao.deleteById(id);
         return String.valueOf(id);
     }
@@ -72,8 +76,8 @@ public class CommentServiceImpl implements CommentService {
         ioService.output("Удалить все комментарии по книге");
         String bookname = ioService.input();
         try {
-            customService.checkBook(bookname);
-            List<Comment> list = checkCommentsByBookName(bookname);
+            checkBook(bookname);
+            List<Comment> list = getCommentsByBookName(bookname);
             list.forEach(temp -> commentDao.deleteById(temp.getId()));
         } catch (IllegalBookException illegalBook) {
             illegalBook.printStackTrace();
@@ -87,16 +91,22 @@ public class CommentServiceImpl implements CommentService {
         ioService.output(String.valueOf(commentDao.count()));
     }
 
-    private Comment checkCommentById(String id) {
+    private Comment getCommentById(String id) {
         Comment temp = commentDao.getById(id);
         if (temp == null) throw new IllegalCommentException();
         return temp;
     }
 
-    private List<Comment> checkCommentsByBookName(String bookname) {
+    private List<Comment> getCommentsByBookName(String bookname) {
         List<Comment> list = commentDao.getByBookName(bookname);
         if (list.size() == 0) throw new IllegalCommentException();
         return list;
+    }
+
+    private Book checkBook(String bookname) throws IllegalBookException {
+        Book temp = bookDao.getByName(bookname);
+        if (temp == null) throw new IllegalBookException(bookname);
+        return temp;
     }
 
 }
